requirements.txt
alembic==1.4.2
blinker==1.4
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
eventlet==0.25.2
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SimpleMDE==0.3.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
markdown2==2.3.9
MarkupSafe==1.1.1
monotonic==1.5
psycopg2==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
